+++ crates/token_proxy_core/src/proxy/upstream.test.rs
@@ -1,3 +1,5 @@
+use axum::http::StatusCode;
+
 use super::utils::is_retryable_status;
 use super::*;
