+++ src/features/config/pages/agents-page.tsx
@@ -1,0 +1,5 @@
+import { ConfigScreen } from "@/features/config/ConfigScreen";
+
+export function AgentsPage() {
+  return <ConfigScreen activeSectionId="agents" />;
+}
