+++ src/features/config/pages/core-page.tsx
@@ -1,0 +1,5 @@
+import { ConfigScreen } from "@/features/config/ConfigScreen";
+
+export function CorePage() {
+  return <ConfigScreen activeSectionId="core" />;
+}
