+++ src/features/config/pages/settings-page.tsx
@@ -1,0 +1,5 @@
+import { ConfigScreen } from "@/features/config/ConfigScreen";
+
+export function SettingsPage() {
+  return <ConfigScreen activeSectionId="settings" />;
+}
