+++ src/features/config/pages/upstreams-page.tsx
@@ -1,0 +1,5 @@
+import { ConfigScreen } from "@/features/config/ConfigScreen";
+
+export function UpstreamsPage() {
+  return <ConfigScreen activeSectionId="upstreams" />;
+}
