+++ src/routes/config/core.lazy.tsx
@@ -1,11 +1,11 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
 
-import { ConfigRoutePage } from "@/features/config/ConfigRoutePage";
+import { CorePage } from "@/features/config/pages/core-page";
 
 export const Route = createLazyFileRoute("/config/core")({
   component: ConfigCoreRoute,
 });
 
 function ConfigCoreRoute() {
-  return <ConfigRoutePage sectionId="core" />;
+  return <CorePage />;
 }
