+++ src/routes/config/logs.lazy.tsx
@@ -1,11 +1,11 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
 
-import { ConfigRoutePage } from "@/features/config/ConfigRoutePage";
+import { LogsPage } from "@/features/logs/pages/logs-page";
 
 export const Route = createLazyFileRoute("/config/logs")({
   component: ConfigLogsRoute,
 });
 
 function ConfigLogsRoute() {
-  return <ConfigRoutePage sectionId="logs" />;
+  return <LogsPage />;
 }
