+++ src/routes/config/upstreams.lazy.tsx
@@ -1,11 +1,11 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
 
-import { ConfigRoutePage } from "@/features/config/ConfigRoutePage";
+import { UpstreamsPage } from "@/features/config/pages/upstreams-page";
 
 export const Route = createLazyFileRoute("/config/upstreams")({
   component: ConfigUpstreamsRoute,
 });
 
 function ConfigUpstreamsRoute() {
-  return <ConfigRoutePage sectionId="upstreams" />;
+  return <UpstreamsPage />;
 }
